:recycle: getDefaultCardList util로 분리
